+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.3
+- Added: The Mythic Dungeon International Returns!
+- Removed: Housing Temporarily Disabled in this Region - Enabled at 7:30 a.m. PDT April 22
+- Removed: Adjustments to Reduce Non-Participation in Dungeons
+
 ## 1.3.2
 - Updated: Raid Tuning Incoming - May 5
 - Updated: Raid Tuning Incoming - 6 May
@@ -130,6 +135,7 @@
 - Added external link copy and guild chat sharing for selected posts.
 - Added login toasts for recent posts.
 - Added class section shortcuts for class hotfix posts.
+